fix some ffeatures, remove mit.js
