+++ includes/listen.js
@@ -66,6 +66,7 @@ module.exports = function({ api, client, __GLOBAL, models, timeStart }) {
 		}
 		catch {
 			if (__GLOBAL.settings.DeveloperMode == true) logger(JSON.stringify(error), "error");
+			else logger("Handle listen đã gặp sự cố: " + error.Error, "eror");
 		}
 	}
 }
