+++ modules/commands/morse.js
@@ -13,6 +13,6 @@ module.exports.config = {
 module.exports.run = function({ api, event, args, utils }) {
     var content = args.join(" ");
     const morsify = require('morsify');
-	if (event.type == "message_reply") (content.indexOf('en') == 0) ? api.sendMessage(morsify.encode(event.messageReply.body), event.threadID, event.messageID) : (content.indexOf('de') == 0) ? api.sendMessage(morsify.decode(event.messageReply.body), event.threadID, event.messageID) : utils.throwError("morse", event.threadID, event.messageID);
-	else (content.indexOf('en') == 0) ? api.sendMessage(morsify.encode(args.slice(1, args.length)), event.threadID, event.messageID) : (content.indexOf('de') == 0) ? api.sendMessage(morsify.decode(args.slice(1, args.length)), event.threadID, event.messageID) : utils.throwError("morse", event.threadID, event.messageID);
+	if (event.type == "message_reply") (content.indexOf('encode') == 0) ? api.sendMessage(morsify.encode(event.messageReply.body), event.threadID, event.messageID) : (content.indexOf('decode') == 0) ? api.sendMessage(morsify.decode(event.messageReply.body), event.threadID, event.messageID) : utils.throwError("morse", event.threadID, event.messageID);
+	else (content.indexOf('encode') == 0) ? api.sendMessage(morsify.encode(`${args.slice(1, args.length)}`), event.threadID, event.messageID) : (content.indexOf('decode') == 0) ? api.sendMessage(morsify.decode(`${args.slice(1, args.length)}`), event.threadID, event.messageID) : utils.throwError("morse", event.threadID, event.messageID);
 }
